+++ src/components/singleElement/Img.ts
