+++ src/components/singleElement/Label.ts
