+++ src/components/singleElement/P.ts
