+++ src/components/singleElement/Span.ts
